Create handler exceptions

+++ src/main/java/br/fjn/pos/api/domain/customers/CustomersService.java
@@ -8,6 +8,7 @@ package br.fjn.pos.api.domain.customers;
 import java.util.List;
 import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
+import javax.ws.rs.NotFoundException;
 
 /**
  *
@@ -15,36 +16,43 @@ import javax.inject.Inject;
  */
 @RequestScoped
 public class CustomersService {
-    
+
     @Inject
     private CustomersRepository customersRepository;
-    
-    public void create(Customer customer) throws Exception{
+
+    public void create(Customer customer) throws Exception {
         this.isValidCPF(customer.getCpf());
         this.customersRepository.create(customer);
     }
-    
-    public void delete(String id){
-      this.customersRepository.delete(id);
+
+    public void delete(String id) {
+        Customer customer = this.customersRepository.findById(id);
+        if (customer == null) {
+            throw new NotFoundException("Entity not found");
+        }
+        this.customersRepository.delete(id);
     }
 
-    public Customer findById(String id){
-      return this.customersRepository.findById(id);
+    public Customer findById(String id) {
+        Customer customer = this.customersRepository.findById(id);
+        if (customer == null) {
+            throw new NotFoundException("Entity not found");
+        }
+        return customer;
     }
 
-    public List<Customer> list(){
-      return this.customersRepository.list();
+    public List<Customer> list() {
+        return this.customersRepository.list();
     }
 
-    public Customer update(Customer customer){
+    public Customer update(Customer customer) {
         return this.customersRepository.update(customer);
     }
-    
-    
-    private void isValidCPF(String cpf) throws Exception{
-        if (cpf.length() != 11){
-            throw new Exception("CPF Inválido");
+
+    private void isValidCPF(String cpf) throws Exception {
+        if (cpf.length() != 11) {
+            throw new Exception("Invalid CPF");
         }
     }
-    
+
 }
